made the Cards and Tasks componization structure
