+++ frontend/src/components/Cards/index.module.css
@@ -1,0 +1,90 @@
+.board_container {
+    display: flex;
+    background-color: var(--color5);
+    height: 78vh;
+    padding: 10px;
+}
+
+.card {
+    background-color: var(--color1);
+    width: 250px;
+    padding: 10px;
+    margin: 5px;
+    border-radius: 5px;
+}
+
+.card_title_container {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.add_card_container {
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+}
+
+.card_title {
+    color: #fff;
+    font-weight: 600;
+    font-size: 12pt;
+    margin: 0 5px;
+}
+
+.task {
+    display: flex;
+    justify-content: space-between;
+    background-color: #fff;
+    padding: 10px;
+    border-radius: 5px;
+    margin: 10px;
+}
+
+.task_title {
+    font-size: 10pt;
+    font-weight: 600;
+    margin: 0 5px;
+    width: 200px;
+}
+
+.add_task_container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 5px;
+    background-color: white;
+    width: 40px;
+}
+
+.add_task_title {
+    font-size: 10pt;
+    font-weight: 600;
+    color: rgb(75, 75, 75);
+}
+
+.icons {
+    color: var(--color-icons);
+    cursor: pointer;
+    font-size: 12pt;
+    margin-left: 5px;
+}
+
+.icon_task_plus {
+    color: var(--color-icons);
+    cursor: pointer;
+    font-size: 12pt;
+    padding: 5px ;
+}
+
+.icon_card_plus {
+    color: white;
+    cursor: pointer;
+    font-size: 12pt;
+    padding: 5px ;
+}
+
+.title_card_icon {
+    color: #fff;
+    font-size: 12pt;
+}
